wrapped the callback exceptions to be thrown back to the client disabled client side checking of permissions to test server side permissions implemented the owner/public data checking for samples/subsamples/chemical analyses there may be some problems with autologin

+++ mpdb-server/src/edu/rpi/metpetdb/server/security/permissions/PermissionInterceptor.java
@@ -63,7 +63,7 @@ public class PermissionInterceptor extends EmptyInterceptor {
 					return;
 				}
 				if (!principals.contains(
-						new OwnerPrincipal(getOwnerId(propertyNames, state)))) {
+						new OwnerPrincipal(getOwnerId(propertyNames, state))) && !isPublic(propertyNames, state)) {
 					throw new CallbackException(
 							"Cannot load objects you don't own, we don't like to share.");
 				}
@@ -170,7 +170,7 @@ public class PermissionInterceptor extends EmptyInterceptor {
 	 */
 	public void onDelete(Object entity, Serializable id, Object[] state,
 			String[] propertyNames, Type[] types) throws CallbackException {
-
+		checkPermissions(entity, state, propertyNames, true);
 		super.onDelete(entity, id, state, propertyNames, types);
 	}
 }
